requirements.txt
aiohttp
rich
python-dotenv
openai